+++ patches/src/main/kotlin/app/morphe/patches/instagram/patches/distractionFree/HideSuggestedContentPatch.kt
@@ -1,52 +1,42 @@
 package app.morphe.patches.instagram.patches.distractionFree
 
-import app.morphe.patcher.Fingerprint
+import app.morphe.patcher.patch.booleanOption
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patches.Constants.COMPATIBILITY_INSTAGRAM
-import app.morphe.patches.instagram.patches.misc.overrideMobileConfigBooleanFlag
-import app.morphe.patches.instagram.utility.replaceJsonFieldWithBogus
-
-private val FEED_ITEM_KEYS_TO_BE_HIDDEN = arrayOf(
-    "clips_netego",
-    "stories_netego",
-    "in_feed_survey",
-    "bloks_netego",
-    "suggested_igd_channels",
-    "suggested_top_accounts",
-    "suggested_users",
-    "suggested_businesses",
-    "suggested_hashtags",
-    "suggested_producers",
-    "suggested_producers_v2",
-    "suggested_close_friends",
-    "suggested_shops"
-)
-
-private object FeedItemParseFromJsonFingerprint : Fingerprint(
-    strings = listOf(*FEED_ITEM_KEYS_TO_BE_HIDDEN, "FeedItem")
-)
+import app.morphe.patches.instagram.misc.overrideMobileConfigBooleanFlag
 
 @Suppress("unused")
-val hideSuggestedContent = bytecodePatch(
+val hideSuggestedContentPatch = bytecodePatch(
     name = "Hide suggested content",
-    description = "Hides suggested stories, reels, threads and survey from feed (Suggested posts will still be shown).",
+    description = "Choose the suggested content to hide using the patch options.",
     default = true
 ) {
     compatibleWith(COMPATIBILITY_INSTAGRAM)
 
+    val hideSuggestedReels by booleanOption(
+        key = "hideSuggestedReels",
+        default = true,
+        title = "Hide suggested reels",
+        description = "Hides suggested reels from feed and reels tab."
+    )
+
+    val hideSuggestedStories by booleanOption(
+        key = "hideSuggestedStories",
+        default = true,
+        title = "Hide suggested stories",
+        description = "Hides suggested stories/users from story tray."
+    )
+
     dependsOn(
         overrideMobileConfigBooleanFlag(
+            // Hides suggestions in search box
             override = "111509::3" to false // ig_search_ta_nullstate_suggestions::is_android_enabled
-        ),
-        overrideMobileConfigBooleanFlag(
-            override = "82771::0" to false // igx_foundation_litho_stories_tray::is_litho_stories_tray_enabled
         )
     )
 
     execute {
-        val matchedMethod = FeedItemParseFromJsonFingerprint.method
-        FEED_ITEM_KEYS_TO_BE_HIDDEN.forEach { key ->
-            matchedMethod.replaceJsonFieldWithBogus(key)
-        }
+        if (hideSuggestedReels == true) hideSuggestedReelsPatch()
+
+        if (hideSuggestedStories == true) hideSuggestedStoriesPatch()
     }
 }